Time: 21 ms (10.52%) | Memory: 39.5 MB (5.02%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -12,7 +12,7 @@ int minEatingSpeed(vector<int>&nums, int h) {
     int low =1; 
     int high = *max_element(nums.begin(), nums.end());
     while(low<=high){
-        int mid = (low+high)/2;
+        long long mid = (low+high)/2;
         long long reqTime = reqHours(nums,mid);
         if(reqTime<=h){
             high = mid-1;
